pyproject.toml
[tool.poetry]
name = "gt-cdt-swe-test-by-jonaw"
version = "0.1.0"
description = "A fictitious Schemes Management System"
authors = ["Jonathan Aw <jonathan.aw.k.h@gmail.com>"]
license = "GNU GENERAL PUBLIC LICENSE Version 3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = "^2.0.32"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
environs = "^11.0.0"
python-dateutil = "^2.9.0.post0"
flask = "^3.0.3"
flask-restful = "^0.3.10"
flask-jwt-extended = "^4.6.0"
marshmallow = "^3.22.0"
marshmallow-sqlalchemy = "^1.1.0"
flask-marshmallow = "^1.2.1"
flask-sqlalchemy = "^3.1.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
